+++ src/ui/textbox.h
@@ -34,7 +34,7 @@ class _TextBox : public _Element {
 		_TextBox();
 		~_TextBox() override;
 
-		void Update(double FrameTime, const glm::ivec2 &Mouse) override;
+		void Update(double FrameTime, const glm::vec2 &Mouse) override;
 		bool HandleKeyEvent(const _KeyEvent &KeyEvent) override;
 		void HandleInput(bool Pressed) override;
 		void Render() const override;
